refactor sc deployer code for reuse, fix errs in logic for deploying mock and non-mock
