+++ sequencer/src/bin/deploy.rs
@@ -1,33 +1,19 @@
-use anyhow::{ensure, Context};
+use anyhow::Context;
 use async_compatibility_layer::logging::{setup_backtrace, setup_logging};
 use async_std::sync::Arc;
-use clap::{builder::OsStr, Parser};
+use clap::Parser;
 use contract_bindings::{
-    erc1967_proxy::ERC1967Proxy,
-    hot_shot::HotShot,
-    light_client::LIGHTCLIENT_ABI,
-    light_client_mock::{LightClientMock, LIGHTCLIENTMOCK_ABI},
-    light_client_state_update_vk_mock::LightClientStateUpdateVKMock,
-    plonk_verifier::PlonkVerifier,
-    shared_types::LightClientState,
+    erc1967_proxy::ERC1967Proxy, hot_shot::HotShot, light_client::LightClient,
 };
-use derive_more::Display;
-use ethers::{
-    contract::Contract as ContractBindings,
-    prelude::{coins_bip39::English, *},
-    solc::artifacts::BytecodeObject,
-};
-use futures::future::{BoxFuture, FutureExt};
-use hotshot_contract_adapter::light_client::ParsedLightClientState;
+use ethers::prelude::{coins_bip39::English, *};
+use futures::future::FutureExt;
 use hotshot_stake_table::config::STAKE_TABLE_CAPACITY;
 use hotshot_state_prover::service::light_client_genesis;
-use std::{
-    collections::HashMap,
-    fs::File,
-    io::{stdout, Write},
-    ops::Deref,
-    path::PathBuf,
+use sequencer::deployer::{
+    deploy_light_client_contract, deploy_mock_light_client_contract, Contract, Contracts,
+    DeployedContracts,
 };
+use std::{fs::File, io::stdout, path::PathBuf};
 use url::Url;
 
 /// Deploy contracts needed to run the sequencer.
@@ -105,131 +91,6 @@ struct Options {
     pub stake_table_capacity: usize,
 }
 
-/// Set of predeployed contracts.
-#[derive(Clone, Debug, Parser)]
-struct DeployedContracts {
-    /// Use an already-deployed HotShot.sol instead of deploying a new one.
-    #[clap(long, env = Contract::HotShot)]
-    hotshot: Option<Address>,
-
-    /// Use an already-deployed PlonkVerifier.sol instead of deploying a new one.
-    #[clap(long, env = Contract::PlonkVerifier)]
-    plonk_verifier: Option<Address>,
-
-    /// Use an already-deployed LightClientStateUpdateVK.sol instead of deploying a new one.
-    #[clap(long, env = Contract::StateUpdateVK)]
-    light_client_state_update_vk: Option<Address>,
-
-    /// Use an already-deployed LightClient.sol instead of deploying a new one.
-    #[clap(long, env = Contract::LightClient)]
-    light_client: Option<Address>,
-
-    /// Use an already-deployed LightClient.sol proxy instead of deploying a new one.
-    #[clap(long, env = Contract::LightClientProxy)]
-    light_client_proxy: Option<Address>,
-}
-
-/// An identifier for a particular contract.
-#[derive(Clone, Copy, Debug, Display, PartialEq, Eq, Hash)]
-enum Contract {
-    #[display(fmt = "ESPRESSO_SEQUENCER_HOTSHOT_ADDRESS")]
-    HotShot,
-    #[display(fmt = "ESPRESSO_SEQUENCER_PLONK_VERIFIER_ADDRESS")]
-    PlonkVerifier,
-    #[display(fmt = "ESPRESSO_SEQUENCER_LIGHT_CLIENT_STATE_UPDATE_VK_ADDRESS")]
-    StateUpdateVK,
-    #[display(fmt = "ESPRESSO_SEQUENCER_LIGHT_CLIENT_ADDRESS")]
-    LightClient,
-    #[display(fmt = "ESPRESSO_SEQUENCER_LIGHT_CLIENT_PROXY_ADDRESS")]
-    LightClientProxy,
-}
-
-impl From<Contract> for OsStr {
-    fn from(c: Contract) -> OsStr {
-        c.to_string().into()
-    }
-}
-
-/// Cache of contracts predeployed or deployed during this current run.
-struct Contracts(HashMap<Contract, Address>);
-
-impl From<DeployedContracts> for Contracts {
-    fn from(deployed: DeployedContracts) -> Self {
-        let mut m = HashMap::new();
-        if let Some(addr) = deployed.hotshot {
-            m.insert(Contract::HotShot, addr);
-        }
-        if let Some(addr) = deployed.plonk_verifier {
-            m.insert(Contract::PlonkVerifier, addr);
-        }
-        if let Some(addr) = deployed.light_client_state_update_vk {
-            m.insert(Contract::StateUpdateVK, addr);
-        }
-        if let Some(addr) = deployed.light_client {
-            m.insert(Contract::LightClient, addr);
-        }
-        if let Some(addr) = deployed.light_client_proxy {
-            m.insert(Contract::LightClientProxy, addr);
-        }
-        Self(m)
-    }
-}
-
-impl Contracts {
-    /// Deploy a contract by calling a function.
-    ///
-    /// The `deploy` function will be called only if contract `name` is not already deployed;
-    /// otherwise this function will just return the predeployed address. The `deploy` function may
-    /// access this [`Contracts`] object, so this can be used to deploy contracts recursively in
-    /// dependency order.
-    async fn deploy_fn(
-        &mut self,
-        name: Contract,
-        deploy: impl FnOnce(&mut Self) -> BoxFuture<'_, anyhow::Result<Address>>,
-    ) -> anyhow::Result<Address> {
-        if let Some(addr) = self.0.get(&name) {
-            tracing::info!("skipping deployment of {name}, already deployed at {addr:#x}");
-            return Ok(*addr);
-        }
-        tracing::info!("deploying {name}");
-        let addr = deploy(self).await?;
-        tracing::info!("deployed {name} at {addr:#x}");
-
-        self.0.insert(name, addr);
-        Ok(addr)
-    }
-
-    /// Deploy a contract by executing its deploy transaction.
-    ///
-    /// The transaction will only be broadcast if contract `name` is not already deployed.
-    async fn deploy_tx<M, C>(
-        &mut self,
-        name: Contract,
-        tx: ContractDeployer<M, C>,
-    ) -> anyhow::Result<Address>
-    where
-        M: Middleware + 'static,
-        C: Deref<Target = ContractBindings<M>> + From<ContractInstance<Arc<M>, M>> + Send + 'static,
-    {
-        self.deploy_fn(name, |_| {
-            async {
-                let contract = tx.send().await?;
-                Ok(contract.address())
-            }
-            .boxed()
-        })
-        .await
-    }
-
-    /// Write a .env file.
-    fn write(&self, mut w: impl Write) -> anyhow::Result<()> {
-        for (contract, address) in &self.0 {
-            writeln!(w, "{contract}={address:#x}")?;
-        }
-        Ok(())
-    }
-}
-
 #[async_std::main]
 async fn main() -> anyhow::Result<()> {
     setup_logging();
@@ -251,38 +112,37 @@ async fn main() -> anyhow::Result<()> {
     contracts
         .deploy_tx(Contract::HotShot, HotShot::deploy(l1.clone(), ())?)
         .await?;
-    contracts
-        .deploy_fn(Contract::LightClientProxy, |contracts| {
-            let l1 = l1.clone();
-            let orchestrator_url = opt.orchestrator_url.clone();
-            async move {
-                let light_client = LightClientMock::new(
-                    contracts
-                        .deploy_fn(Contract::LightClient, |contracts| {
-                            deploy_light_client_contract(
-                                l1.clone(),
-                                contracts,
-                                opt.use_mock_contract,
-                            )
-                            .boxed()
-                        })
-                        .await?,
-                    l1.clone(),
-                );
-                let genesis =
-                    light_client_genesis(&orchestrator_url, opt.stake_table_capacity).await?;
-                let data = light_client
-                    .initialize(genesis.into(), u32::MAX, owner)
-                    .calldata()
-                    .context("calldata for initialize transaction not available")?;
-                let proxy = ERC1967Proxy::deploy(l1, (light_client.address(), data))?
-                    .send()
-                    .await?;
-                Ok(proxy.address())
-            }
-            .boxed()
-        })
-        .await?;
+
+    if opt.use_mock_contract {
+        // LightClientMock is a non-upgradable contract, thus directly initialize
+        // it via its constructor
+        contracts
+            .deploy_fn(Contract::LightClient, |contracts| {
+                deploy_mock_light_client_contract(l1.clone(), contracts, None).boxed()
+            })
+            .await?;
+    } else {
+        // LightClient is a upgradable contract, thus deploy first,
+        // then initialize it through a proxy contract
+        let lc_address = contracts
+            .deploy_fn(Contract::LightClient, |contracts| {
+                deploy_light_client_contract(l1.clone(), contracts).boxed()
+            })
+            .await?;
+        let light_client = LightClient::new(lc_address, l1.clone());
+
+        let genesis = light_client_genesis(&opt.orchestrator_url, opt.stake_table_capacity).await?;
+        let data = light_client
+            .initialize(genesis.into(), u32::MAX, owner)
+            .calldata()
+            .context("calldata for initialize transaction not available")?;
+        contracts
+            .deploy_tx(
+                Contract::LightClientProxy,
+                ERC1967Proxy::deploy(l1.clone(), (lc_address, data))?,
+            )
+            .await?;
+    }
 
     if let Some(out) = &opt.out {
         let file = File::options().create(true).write(true).open(out)?;
@@ -292,119 +152,4 @@ async fn main() -> anyhow::Result<()> {
     }
 
     Ok(())
-}
-
-async fn deploy_light_client_contract<M: Middleware + 'static>(
-    l1: Arc<M>,
-    contracts: &mut Contracts,
-    use_mock_contract: bool,
-) -> anyhow::Result<Address> {
-    // Deploy library contracts.
-    let plonk_verifier = contracts
-        .deploy_tx(
-            Contract::PlonkVerifier,
-            PlonkVerifier::deploy(l1.clone(), ())?,
-        )
-        .await?;
-    let vk = contracts
-        .deploy_tx(
-            Contract::StateUpdateVK,
-            LightClientStateUpdateVKMock::deploy(l1.clone(), ())?,
-        )
-        .await?;
-    if use_mock_contract {
-        deploy_mock_light_client_contract(l1, plonk_verifier, vk).await
-    } else {
-        deploy_production_light_client_contract(l1, plonk_verifier, vk).await
-    }
-}
-
-async fn deploy_production_light_client_contract<M: Middleware + 'static>(
-    l1: Arc<M>,
-    plonk_verifier: Address,
-    vk: Address,
-) -> anyhow::Result<Address> {
-    // Link with LightClient's bytecode artifacts. We include the unlinked bytecode for the contract
-    // in this binary so that the contract artifacts do not have to be distributed with the binary.
-    // This should be fine because if the bindings we are importing are up to date, so should be the
-    // contract artifacts: this is no different than foundry inlining bytecode objects in generated
-    // bindings, except that foundry doesn't provide the bytecode for contracts that link with
-    // libraries, so we have to do it ourselves.
-    let mut bytecode: BytecodeObject = serde_json::from_str(include_str!(
-        "../../../contract-bindings/artifacts/LightClient_bytecode.json",
-    ))?;
-    bytecode
-        .link_fully_qualified(
-            "contracts/src/libraries/PlonkVerifier.sol:PlonkVerifier",
-            plonk_verifier,
-        )
-        .resolve()
-        .context("error linking PlonkVerifier lib")?;
-    bytecode
-        .link_fully_qualified(
-            "contracts/tests/mocks/LightClientStateUpdateVK.sol:LightClientStateUpdateVK",
-            vk,
-        )
-        .resolve()
-        .context("error linking LightClientStateUpdateVK lib")?;
-    ensure!(!bytecode.is_unlinked(), "failed to link LightClient.sol");
-
-    // Deploy light client.
-    let light_client_factory = ContractFactory::new(
-        LIGHTCLIENT_ABI.clone(),
-        bytecode
-            .as_bytes()
-            .context("error parsing bytecode for linked LightClient contract")?
-            .clone(),
-        l1,
-    );
-    let contract = light_client_factory.deploy(())?.send().await?;
-    Ok(contract.address())
-}
-
-async fn deploy_mock_light_client_contract<M: Middleware + 'static>(
-    l1: Arc<M>,
-    plonk_verifier: Address,
-    vk: Address,
-) -> anyhow::Result<Address> {
-    // Link with LightClient's bytecode artifacts. We include the unlinked bytecode for the contract
-    // in this binary so that the contract artifacts do not have to be distributed with the binary.
-    // This should be fine because if the bindings we are importing are up to date, so should be the
-    // contract artifacts: this is no different than foundry inlining bytecode objects in generated
-    // bindings, except that foundry doesn't provide the bytecode for contracts that link with
-    // libraries, so we have to do it ourselves.
-    let mut bytecode: BytecodeObject = serde_json::from_str(include_str!(
-        "../../../contract-bindings/artifacts/LightClientMock_bytecode.json",
-    ))?;
-    bytecode
-        .link_fully_qualified(
-            "contracts/src/libraries/PlonkVerifier.sol:PlonkVerifier",
-            plonk_verifier,
-        )
-        .resolve()
-        .context("error linking PlonkVerifier lib")?;
-    bytecode
-        .link_fully_qualified(
-            "contracts/tests/mocks/LightClientStateUpdateVKMock.sol:LightClientStateUpdateVKMock",
-            vk,
-        )
-        .resolve()
-        .context("error linking LightClientStateUpdateVK lib")?;
-    ensure!(!bytecode.is_unlinked(), "failed to link LightClient.sol");
-
-    // Deploy light client.
-    let light_client_factory = ContractFactory::new(
-        LIGHTCLIENTMOCK_ABI.clone(),
-        bytecode
-            .as_bytes()
-            .context("error parsing bytecode for linked LightClient contract")?
-            .clone(),
-        l1,
-    );
-    let dummy_genesis: LightClientState = ParsedLightClientState::dummy_genesis().into();
-    let contract = light_client_factory
-        .deploy((dummy_genesis, u32::MAX))?
-        .send()
-        .await?;
-    Ok(contract.address())
 }
