+++ sequencer/src/lib.rs
@@ -3,6 +3,7 @@ pub mod block;
 pub mod catchup;
 mod chain_variables;
 pub mod context;
+pub mod deployer;
 mod header;
 pub mod hotshot_commitment;
 pub mod options;
